moved maze_generator.py to src folder and created user_interaction.py file in src
